# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tractor Work Tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  mobile: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  mobile: string;
  role: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface CreateUserBody {
  name: string;
  mobile: string;
  password: string;
  role: string;
}

export interface UpdateUserBody {
  password: string;
  /** @nullable */
  name?: string | null;
}

export interface Worker {
  id: number;
  name: string;
  /** @nullable */
  mobile?: string | null;
  createdAt: string;
}

export interface WorkerSummary {
  id: number;
  name: string;
  /** @nullable */
  mobile?: string | null;
  createdAt: string;
  totalEarned: number;
  totalPaid: number;
  pendingAmount: number;
}

export interface CreateWorkerBody {
  name: string;
  mobile: string;
}

export type WorkerConflictResponseExistingWorker = {
  id: number;
  name: string;
  mobile?: string;
};

export interface WorkerConflictResponse {
  error: string;
  existingWorker: WorkerConflictResponseExistingWorker;
}

export type WorkLogStatus = (typeof WorkLogStatus)[keyof typeof WorkLogStatus];

export const WorkLogStatus = {
  PENDING: "PENDING",
  PARTIAL: "PARTIAL",
  PAID: "PAID",
} as const;

export interface WorkLog {
  id: number;
  workerId: number;
  /** @nullable */
  workerName?: string | null;
  fieldName: string;
  startTime: string;
  endTime: string;
  totalHours: number;
  amount: number;
  paidAmount: number;
  status: WorkLogStatus;
  /** @nullable */
  jcbUserId?: number | null;
  createdAt: string;
}

export interface CreateWorkLogBody {
  workerId: number;
  fieldName: string;
  startTime: string;
  endTime: string;
  totalHours: number;
  amount: number;
}

export interface ActiveWorkLog {
  active: boolean;
  /** @nullable */
  workerId?: number | null;
  /** @nullable */
  workerName?: string | null;
  /** @nullable */
  fieldName?: string | null;
  /** @nullable */
  startTime?: string | null;
}

export interface Payment {
  id: number;
  workerId: number;
  amountPaid: number;
  paymentDate: string;
}

export interface CreatePaymentBody {
  workerId: number;
  amountPaid: number;
  jcbUserId?: number | null;
}

export interface RateSetting {
  amountPerHour: number;
}

export interface SetRateBody {
  amountPerHour: number;
}

export interface DashboardSummary {
  totalWorkers: number;
  totalEarned: number;
  totalPaid: number;
  totalPending: number;
  amountPerHour: number;
}