# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tractor Work Tracker
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with mobile + password
 */
export const LoginBody = zod.object({
  mobile: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    mobile: zod.string(),
    role: zod.string(),
    createdAt: zod.string(),
  }),
});

/**
 * @summary Get current user info
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  mobile: zod.string(),
  role: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary List all users (admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  mobile: zod.string(),
  role: zod.string(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a new user (admin only)
 */
export const CreateUserBody = zod.object({
  name: zod.string(),
  mobile: zod.string(),
  password: zod.string(),
  role: zod.string(),
});

/**
 * @summary Update user password (admin only)
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  password: zod.string(),
  name: zod.string().nullish(),
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  mobile: zod.string(),
  role: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a user (admin only)
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all workers with payment summary
 */
export const ListWorkersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  mobile: zod.string().nullish(),
  createdAt: zod.string(),
  totalEarned: zod.number(),
  totalPaid: zod.number(),
  pendingAmount: zod.number(),
});
export const ListWorkersResponse = zod.array(ListWorkersResponseItem);

/**
 * @summary Create a new worker
 */
export const CreateWorkerBody = zod.object({
  name: zod.string(),
  mobile: zod.string(),
});

/**
 * @summary Get single worker details
 */
export const GetWorkerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWorkerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  mobile: zod.string().nullish(),
  createdAt: zod.string(),
  totalEarned: zod.number(),
  totalPaid: zod.number(),
  pendingAmount: zod.number(),
});

/**
 * @summary Get work logs for a worker
 */
export const GetWorkerLogsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWorkerLogsResponseItem = zod.object({
  id: zod.number(),
  workerId: zod.number(),
  workerName: zod.string().nullish(),
  fieldName: zod.string(),
  startTime: zod.string(),
  endTime: zod.string(),
  totalHours: zod.number(),
  amount: zod.number(),
  paidAmount: zod.number(),
  status: zod.enum(["PENDING", "PARTIAL", "PAID"]),
  jcbUserId: zod.number().nullish(),
  createdAt: zod.string(),
});
export const GetWorkerLogsResponse = zod.array(GetWorkerLogsResponseItem);

/**
 * @summary Get payment history for a worker
 */
export const GetWorkerPaymentsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWorkerPaymentsResponseItem = zod.object({
  id: zod.number(),
  workerId: zod.number(),
  amountPaid: zod.number(),
  paymentDate: zod.string(),
});
export const GetWorkerPaymentsResponse = zod.array(
  GetWorkerPaymentsResponseItem,
);

/**
 * @summary Save a completed work session
 */
export const CreateWorkLogBody = zod.object({
  workerId: zod.number(),
  fieldName: zod.string(),
  startTime: zod.string(),
  endTime: zod.string(),
  totalHours: zod.number(),
  amount: zod.number(),
});

/**
 * @summary Get active (in-progress) work session if any
 */
export const GetActiveWorkLogResponse = zod.object({
  active: zod.boolean(),
  workerId: zod.number().nullish(),
  workerName: zod.string().nullish(),
  fieldName: zod.string().nullish(),
  startTime: zod.string().nullish(),
});

/**
 * @summary Add a payment for a worker
 */
export const CreatePaymentBody = zod.object({
  workerId: zod.number(),
  amountPaid: zod.number(),
  jcbUserId: zod.number().nullish(),
});

/**
 * @summary Get current amount per hour
 */
export const GetRateResponse = zod.object({
  amountPerHour: zod.number(),
});

/**
 * @summary Set amount per hour
 */
export const SetRateBody = zod.object({
  amountPerHour: zod.number(),
});

export const SetRateResponse = zod.object({
  amountPerHour: zod.number(),
});

/**
 * @summary Get overall dashboard summary totals
 */
export const GetDashboardSummaryResponse = zod.object({
  totalWorkers: zod.number(),
  totalEarned: zod.number(),
  totalPaid: zod.number(),
  totalPending: zod.number(),
  amountPerHour: zod.number(),
});