# lib/api-zod/src/generated/types/setRateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tractor Work Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface SetRateBody {
  amountPerHour: number;
}