# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tractor Work Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  mobile: string;
  role: string;
  createdAt: string;
}